Fixed bug in Schedule Summary page where not all projects could be viewed in project report
Not sure how the original bug came about, but when no project was entered in
the project text field, we returned a blank report - well, that was not how
it *used* to behave.  Now it should give a project report for all projects
when the project text field is left blank.